Refactor RustPlus API to use ulong for entity IDs and improve response handling
- Updated SetSmartSwitchValueAsync, SetSubscriptionAsync, StrobeSmartSwitchAsync, and ToggleSmartSwitchAsync methods to use ulong instead of uint for entity IDs.
- Modified Program.cs to change ReadLine to ReadKey for better user interaction.
- Enhanced DisplayUtilities to handle Response objects more effectively.
- Introduced RustPlusFcmSocketOptions for configurable heartbeat and inactivity timeout settings.
- Refactored IRustPlus interface methods to return a payload-free Response for acknowledge-only commands.
- Updated tests to reflect changes in response handling and entity ID types.
- Improved error handling in GetErrorMessage method to return "unknown-error" for non-error responses.

+++ samples/RustPlus.ConsoleApp/Features/CheckSubscription.cs
@@ -5,7 +5,7 @@ namespace RustPlus.ConsoleApp.Features;
 
 internal sealed class CheckSubscription(IRustPlus rustPlus)
 {
-    public async Task CheckSubscriptionAsync(uint alarmId)
+    public async Task CheckSubscriptionAsync(ulong alarmId)
     {
         var response = await rustPlus.CheckSubscriptionAsync(alarmId);
         DisplayUtilities.DisplayJson("AlarmInfo", response);
