+++ samples/RustPlus.ConsoleApp/Features/GetAlarmInfo.cs
@@ -5,7 +5,7 @@ namespace RustPlus.ConsoleApp.Features;
 
 internal sealed class GetAlarmInfo(IRustPlus rustPlus)
 {
-    public async Task GetAlarmInfoAsync(uint entityId)
+    public async Task GetAlarmInfoAsync(ulong entityId)
     {
         var response = await rustPlus.GetAlarmInfoAsync(entityId);
         DisplayUtilities.DisplayJson("AlarmInfo", response);
