+++ samples/RustPlus.ConsoleApp/Features/GetStorageMonitorInfo.cs
@@ -5,7 +5,7 @@ namespace RustPlus.ConsoleApp.Features;
 
 internal sealed class GetStorageMonitorInfo(IRustPlus rustPlus)
 {
-    public async Task GetStorageMonitorInfoAsync(uint entityId)
+    public async Task GetStorageMonitorInfoAsync(ulong entityId)
     {
         var response = await rustPlus.GetStorageMonitorInfoAsync(entityId);
         DisplayUtilities.DisplayJson("StorageMonitorInfo", response);
