+++ samples/RustPlus.ConsoleApp/Features/SetSmartSwitchValue.cs
@@ -5,7 +5,7 @@ namespace RustPlus.ConsoleApp.Features;
 
 internal sealed class SetSmartSwitchValue(IRustPlus rustPlus)
 {
-    public async Task SetSmartSwitchValueAsync(uint smartSwitchId, bool smartSwitchValue)
+    public async Task SetSmartSwitchValueAsync(ulong smartSwitchId, bool smartSwitchValue)
     {
         var response = await rustPlus.SetSmartSwitchValueAsync(smartSwitchId, smartSwitchValue);
